docker: Add initial release building dockerfiles
There are different dockerfiles for each release configuration. They
have a lot in common, but docker does not have a built in include
syntax, and I didn't feel like it was worth adding templating at the
present level of complexity. If someone felt strongly about this or we
started increasing the number of targets the files could probably be
templated straightforwardly to reduce a lot of the duplication.