do_build: Manually implement copytree for deb synchronization
The copytree method will raise an error by default if the destination
directory already exists, instead of just merging in the new
files. Starting in python 3.8 there is a parameter to ignore this
behavior, but some of our build hosts have older version of
python. This commit adds a temporary manual implementation to copy the
files over without throwing an error on parent directories, and which
works back to python 3.5.